Translation/requirements.txt
# Google Cloud Translation API package required
google-cloud-translate
